New hack: train ML and PG together
This way we ensure we don't forget ML's teachings

I think the problem is with my method of PG training: I'm only ever
giving examples with several words already output.
The Policy Gradient never gets a chance to optimize the earliest
words of a sentence, which is where it might be needed the most.